+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added ResourceTemplate resource. ResourceTemplate will be used to populate
 namespaces with initial resources.
 - Add `EntityServiceClass` constant to the `corev2` package, representing BSM Services.
+- Added ResourceTemplate instantiation on namespace creation.
 
 ### Fixed
 - Both V2 & V3 resources are now validated when used with storev2.